Removed jest from craco and moved resolveAsBin
