+++ packages/modular-scripts/package.json
@@ -13,7 +13,7 @@
     "modular": "ts-node src/cli.ts",
     "clean": "rimraf build",
     "prebuild": "yarn clean",
-    "build": "babel --root-mode upward src --out-dir build --extensions .ts,.js --ignore 'src/**/*.test.ts'"
+    "build": "babel --root-mode upward src --out-dir build --extensions .ts --ignore 'src/**/*.test.ts'"
   },
   "dependencies": {
     "@babel/core": "7.14.0",
