+++ packages/modular-scripts/src/start.ts
@@ -1,9 +1,10 @@
 import * as path from 'path';
+import resolveAsBin from 'resolve-as-bin';
 import getModularRoot from './utils/getModularRoot';
 import isModularType from './utils/isModularType';
 import execSync from './utils/execSync';
 
-import { cracoBin, packagesRoot, cracoConfig } from './config';
+import { packagesRoot, cracoConfig } from './config';
 
 export default async function start(appPath: string): Promise<void> {
   const modularRoot = getModularRoot();
@@ -11,6 +12,7 @@ export default async function start(appPath: string): Promise<void> {
   if (!isModularType(path.join(modularRoot, packagesRoot, appPath), 'app')) {
     throw new Error(`The package at ${appPath} is not a valid modular app.`);
   }
+  const cracoBin = resolveAsBin('craco');
 
   execSync(cracoBin, ['start', '--config', cracoConfig], {
     cwd: path.join(modularRoot, packagesRoot, appPath),
